// mobileFrontend/src/shared/gen/ts/proto/auth_service.ts
export interface LoginRequest {
  /**
   * @generated from protobuf field: string username = 1;
   */
  username: string;
  /**
   * @generated from protobuf field: string password = 2;
   */
  password: string;
}
/**
 * @generated from protobuf message pb.LoginResponse
 */
export interface LoginResponse {
  /**
   * @generated from protobuf field: string token = 1;
   */
  token: string;
}
// ...
class LoginRequest$Type extends MessageType<LoginRequest> {
  constructor() {
    super("pb.LoginRequest", [
      {
        no: 1,
        name: "username",
        kind: "scalar",
        T: 9 /*ScalarType.STRING*/,
        options: { "validate.rules": { string: { email: true } } },
      },
      { no: 2, name: "password", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
    ]);
  }
  create(value?: PartialMessage<LoginRequest>): LoginRequest {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.username = "";
    message.password = "";
    if (value !== undefined)
      reflectionMergePartial<LoginRequest>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: LoginRequest,
  ): LoginRequest {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* string username */ 1:
          message.username = reader.string();
          break;
        case /* string password */ 2:
          message.password = reader.string();
          break;
        default:
          let u = options.readUnknownField;
          if (u === "throw")
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: LoginRequest,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* string username = 1; */
    if (message.username !== "")
      writer.tag(1, WireType.LengthDelimited).string(message.username);
    /* string password = 2; */
    if (message.password !== "")
      writer.tag(2, WireType.LengthDelimited).string(message.password);
    let u = options.writeUnknownFields;
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message pb.LoginRequest
 */
export const LoginRequest = new LoginRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class LoginResponse$Type extends MessageType<LoginResponse> {
  constructor() {
    super("pb.LoginResponse", [
      { no: 1, name: "token", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
    ]);
  }
  create(value?: PartialMessage<LoginResponse>): LoginResponse {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.token = "";
    if (value !== undefined)
      reflectionMergePartial<LoginResponse>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: LoginResponse,
  ): LoginResponse {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* string token */ 1:
          message.token = reader.string();
          break;
        default:
          let u = options.readUnknownField;
          if (u === "throw")
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: LoginResponse,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* string token = 1; */
    if (message.token !== "")
      writer.tag(1, WireType.LengthDelimited).string(message.token);
    let u = options.writeUnknownFields;
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message pb.LoginResponse
 */
export const LoginResponse = new LoginResponse$Type();